requirements.txt
fastapi
uvicorn
line-bot-sdk>=3.0.0
httpx
python-dotenv
pydantic
